+++ CHANGES.md
@@ -5,5 +5,6 @@ Change Log
 
 * changed asImage to toDataURL
 * changed asCanvas(ratio: Double = 1) to take a "window" ratio value, default=1.
+* changed all strokeWidth to be Double
 
 * added Outline ---> not yet finished need more info